updated maps and proofs
