+++ clients/vault/src/oracle/storage/impls.rs
@@ -1,14 +1,14 @@
 use std::{fmt::Write, str::Split};
 
 use stellar_relay_lib::sdk::{
-    compound_types::UnlimitedVarArray,
-    types::{ScpEnvelope, ScpHistoryEntry, TransactionHistoryEntry, TransactionSet},
-    XdrCodec,
+	compound_types::UnlimitedVarArray,
+	types::{ScpEnvelope, ScpHistoryEntry, TransactionHistoryEntry, TransactionSet},
+	XdrCodec,
 };
 
 use crate::oracle::{
-    EnvelopesFileHandler, EnvelopesMap, Error, Filename, SerializedData, Slot, SlotEncodedMap,
-    storage::traits::*, TxSetMap, TxSetsFileHandler,
+	storage::traits::*, types::AddExt, EnvelopesFileHandler, EnvelopesMap, Error, Filename,
+	SerializedData, Slot, SlotEncodedMap, TxSetMap, TxSetsFileHandler,
 };
 
 use super::{ScpArchiveStorage, TransactionsArchiveStorage};
@@ -116,7 +116,7 @@ impl FileHandlerExt<TxSetMap> for TxSetsFileHandler {
 				let _ = write!(filename, "{}", key);
 			}
 
-			m.insert(*key, set.to_xdr());
+			m.insert(*key, set.clone().into_bytes());
 		}
 
 		Ok((filename, bincode::serialize(&m)?))
@@ -146,11 +146,11 @@ impl ArchiveStorage for TransactionsArchiveStorage {
 #[cfg(test)]
 mod test {
 	use std::{
-        convert::{TryFrom, TryInto},
-        fs,
-        fs::File,
-        io::Read,
-        path::PathBuf,
+		convert::{TryFrom, TryInto},
+		fs,
+		fs::File,
+		io::Read,
+		path::PathBuf,
 	};
 
 	use mockall::lazy_static;
@@ -158,19 +158,18 @@ mod test {
 	use stellar_relay_lib::sdk::types::ScpHistoryEntry;
 
 	use crate::oracle::{
-        errors::Error,
-        get_test_stellar_relay_config,
-        impls::ArchiveStorage,
-        storage::{
-            EnvelopesFileHandler,
-            traits::{FileHandler, FileHandlerExt}, TxSetsFileHandler,
-        },
-        TransactionsArchiveStorage,
-        types::Slot,
+		errors::Error,
+		get_test_stellar_relay_config,
+		impls::ArchiveStorage,
+		storage::{
+			traits::{FileHandler, FileHandlerExt},
+			EnvelopesFileHandler, TxSetsFileHandler,
+		},
+		types::{constants::MAX_SLOTS_PER_FILE, Slot},
+		TransactionsArchiveStorage,
 	};
-    use crate::oracle::types::constants::MAX_SLOTS_PER_FILE;
 
-    use super::ScpArchiveStorage;
+	use super::ScpArchiveStorage;
 
 	impl Default for ScpArchiveStorage {
 		fn default() -> Self {
