+++ clients/vault/src/oracle/storage/traits.rs
@@ -1,9 +1,9 @@
 use std::{
-    fs,
-    fs::File,
-    io::{Read, Write},
-    path::PathBuf,
-    str::Split,
+	fs,
+	fs::File,
+	io::{Read, Write},
+	path::PathBuf,
+	str::Split,
 };
 
 use flate2::bufread::GzDecoder;
@@ -11,8 +11,9 @@ use sp_core::hexdisplay::AsBytesRef;
 
 use stellar_relay_lib::sdk::{compound_types::XdrArchive, XdrCodec};
 
-use crate::oracle::{Error, Filename, SerializedData, Slot};
-use crate::oracle::types::constants::ARCHIVE_NODE_LEDGER_BATCH;
+use crate::oracle::{
+	types::constants::ARCHIVE_NODE_LEDGER_BATCH, Error, Filename, SerializedData, Slot,
+};
 
 pub trait FileHandlerExt<T: Default>: FileHandler<T> {
 	fn create_filename_and_data(data: &T) -> Result<(Filename, SerializedData), Error>;
