+++ clients/vault/src/oracle/types/mod.rs
@@ -1,8 +1,10 @@
-mod types;
+pub mod constants;
+mod double_sided_map;
 mod limited_fifo_map;
 mod transaction_set;
-pub mod constants;
+mod types;
 
+pub use double_sided_map::*;
+pub use limited_fifo_map::*;
+pub(crate) use transaction_set::*;
 pub use types::*;
-// use limited_fifo_map::*;
-use transaction_set::*;
